+++ Package.swift
@@ -10,7 +10,10 @@ let package = Package(
         .executable(name: "Beadazzle", targets: ["Beadazzle"])
     ],
     dependencies: [
-        .package(url: "https://github.com/nodes-app/swift-markdown-engine", exact: "0.8.0"),
+        .package(
+            url: "https://github.com/Mosnar/swift-markdown-engine",
+            revision: "a2dc2ea551cf65d4c72a1062fb20a7ece0c8cca6"
+        ),
         .package(url: "https://github.com/sparkle-project/Sparkle", exact: "2.9.4")
     ],
     targets: [